Several JavaScript changes to add more interactivity to the form elements
Management of conflicting checkbox input elements so that certain
checkboxes are disabled as others become selected.

Disabled checkbox text color is styled.

Checkboxes are not disabled when conflicting checkbox is deselected by
the user, e.g. user clicks conflicting box and then clicks it again to
mark the box not checked.

totalCost is counting the amount of money to be charged to the user
upon successful check out based on interaction with a (set of) checkbox
element(s).