creditCardValidator plugin added as well as other error message improvements
creditCardValidator plugin added as well as other error message
improvements

Issue seems to arise using non-email errors to check for all page
errors within the validateForm() function of the main.js file. Might it
be from (improper) use of the test() method?

Where to use preventDefault() to stop form submission when there are
errors and if/when to use reset() and/or submit() for when the forms
are all valid?